requirements-container.txt
# Python 3.10 / ROS Humble devcontainer requirements for Empower.
# Torch is intentionally not pinned here; the devcontainer owns the Torch/CUDA
# stack (currently installed from .devcontainer/requirements.txt).

einops
opencv-python
timm>=1.0.17
tqdm>=4.66.1
torchprofile
matplotlib
ultralytics
git+https://github.com/ultralytics/CLIP.git
transformers
onnx
onnxsim
onnxruntime
git+https://github.com/zhijian-liu/torchpack.git@3a5a9f7ac665444e1eb45942ee3f8fc7ffbd84e5
git+https://github.com/alibaba/TinyNeuralNetwork.git
git+https://github.com/facebookresearch/segment-anything.git
pycocotools
lvis
prettytable
terminaltables
openai
requests>=2.31
rich>=13.8
langchain-core
langchain-openai
langchain-mistralai
pyyaml
spacy>=3.7,<3.8
gensim
open3d
scikit-learn
numpy>=1.24,<2
tomli; python_version < "3.11"
